moved dynamic fee config to fee package
