+++ vms/platformvm/txs/fee/dynamic_config.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package config
+package fee
 
 import (
 	"fmt"
